Code cleanup, 2

+++ src/musicxml/mxmlTree2MsrTranslator.h
@@ -1064,7 +1064,7 @@ class mxmlTree2MsrTranslator :
     // ------------------------------------------------------
 
     float                     fCurrentMillimeters;
-    int                       fCurrentTenths;
+    float                     fCurrentTenths;
 
 
     // page layout handling
